model_zoo/requirements.txt
datasets #==3.0.0
einops #==0.8.0
loralib #==0.1.2
matplotlib #==3.9.2
numpy #==2.1.1
pandas #==2.2.2
peft
Pillow #==10.4.0
pyarrow #==17.0.0
PyYAML #==6.0.2
Requests #==2.32.3
schedulefree #==1.2.7
scipy #==1.14.1
timm #==1.0.9
torch #==2.4.1
torchvision #==0.19.1
tqdm #==4.66.5
train #==0.0.5
wandb #==0.18.0
xtcocotools #==1.14.3
